% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_test.R
\name{agg_test}
\alias{agg_test}
\title{Aggregation testing using RAREMETAL and external covariance}
\usage{
agg_test(
  score_stat_file,
  vcf_file,
  chr,
  burden = TRUE,
  wburden = FALSE,
  SKAT = FALSE,
  anno_file = NULL,
  anno = NULL,
  two_stage = FALSE,
  two_stage_threshold = 3,
  group_file = NULL,
  pLOF = FALSE,
  pLOF_narrowMissense = FALSE,
  pLOF_broadMissense = FALSE,
  altGroupFilePath = NULL,
  altCovariancePath = NULL,
  altRaremetalPath = NULL,
  mafThreshold = 0.01,
  gene = NULL,
  gene_start = NULL,
  gene_end = NULL,
  hwe = 1e-06
)
}
\arguments{
\item{score_stat_file}{String containing the file path and name containing
the RAREMETALWORKER formatted score statistic file.}

\item{vcf_file}{String containing the file path and name containing the VCF
of the genotypes of the  set of individuals to be used as the reference
panel.}

\item{chr}{The chromosome number on which you want to perform the analysis.
Must be in the same format as in the CHROM column of the VCF. This could be
for example c{chr_number}, chr{chr_number}, or {chr_number}. ie. c1, chr1, or 1
for chromosome 1.}

\item{burden}{Perform simple (unweighted) burden test? Must be TRUE or FALSE.
Default = TRUE.}

\item{wburden}{Perform weight burden test? Weights are from beta(MAF, 1, 25)
distribution. Must be TRUE or FALSE. Default = FALSE.}

\item{SKAT}{Perform simple (unweighted) SKAT? Must be TRUE or FALSE.
Default = FALSE.}

\item{anno_file}{String containing the file path and name containing the
VEP annotation file to create the group files. If not specified, must give
argument to group_file. Default = NULL.}

\item{anno}{Option to directly pass VEP annotation file to the function.
Default = NULL. Do not pass both anno_file and anno to the function as it will
cause an error.}

\item{two_stage}{TRUE or FALSE. Do you want to perform two-stage method
where external reference panel is used only when the -log10(p-value) using
the null covariance is above some threshold? Default = FALSE.}

\item{two_stage_threshold}{What -log10(p-value) threshold would you like to
use for two-stage approach? Default = 3.}

\item{group_file}{Vector of strings containing the file paths and names of
the group files to perform the rare-variant aggregation tests with. If not
specified, must supply anno_file (or ) and types of group files to use. Default =
NULL.}

\item{pLOF}{TRUE or FALSE. Do you want to perform aggregation tests using
pLOF variants only? Default = FALSE.}

\item{pLOF_narrowMissense}{TRUE or FALSE. Do you want to perform aggregation
tests using pLOF + missense(narrow) variants only? Default = FALSE.}

\item{pLOF_broadMissense}{TRUE or FALSE. Do you want to perform aggregation
tests using pLOF + missense(broad) variants only? Default = FALSE.}

\item{altGroupFilePath}{If pLOF, pLOF_narrowMissense, or pLOF_broadMissense
are TRUE, option to give alternative path to write group files to. Default =
NULL. If left as NULL, group files will be written in current directory.}

\item{altCovariancePath}{Option to give alternative path to write covariance
files to. Default = NULL. If left as NULL, covariance files will be written
in current directory.}

\item{altRaremetalPath}{Optional alternative path to write RAREMETAL results
to. Default = NULL. If left as NULL, will write in current directory.}

\item{mafThreshold}{If pLOF, pLOF_narrowMissense, or pLOF_broadMissense
are TRUE, threshold for including variants in group file. Note here
we assume that the alternate allele is coded as the minor allele.
Default = 0.01.}

\item{gene}{The name of the gene to use for annotation files if group files
are not already supplied. Default = NULL.}

\item{gene_start}{The starting base pair position of the gene on the chromosome.
Default = NULL. If you have large reference panel VCFs it is recommended to
supply gene_start and gene_end.}

\item{gene_end}{The ending base pair position of the gene on the chromosome.
Default = NULL. If you have large reference panel VCFs it is recommended to
supply gene_start and gene_end.}

\item{hwe}{Hardy-Weinberg equilibrium p-value cut-off for rare-variant
aggregation testing. Default = 0.000001.}
}
\value{
1 if RAREMETAL was performed using the estimated covariance successfully,
0 if two-stage approach was performed successfully but covariance did not need
to be computed.
}
\description{
A function that perfroms various rare-variant aggregation tests
 given a path to a RAREMETALWORKER formatted score statistic file, a path to
 a VCF containing the individual-level genetic data of the individuals to be
 used as the external reference panel, and the path to a gzipped VEP
 annotation file. Option to use the two-stage approach where external
 reference panel is used only when the -log10(p-value) using the null
 covariance is above some threshold.
}
\examples{
#Use examples of data that I simulate (Note this will cause devtool::check() issue)
}
