% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_raremetal.R
\name{call_raremetal}
\alias{call_raremetal}
\title{Burden testing using RAREMETAL}
\usage{
call_raremetal(
  mask_list,
  score_stat_file,
  group_file_names,
  burden,
  wburden,
  SKAT,
  altCovariancePath = NULL,
  altGroupFilePath = NULL,
  altRaremetalPath = NULL,
  gene = "gene",
  hwe = 1e-06,
  altRaremetalName = NULL
)
}
\arguments{
\item{mask_list}{Named list of vectors containing the variants to use in the
group files for burden testing.}

\item{score_stat_file}{String containing the file path and name containing
the RAREMETALWORKER formatted score statistic file.}

\item{group_file_names}{Vector of strings containing the file paths and names of
the group files to perform the rare-variant aggregation tests with.}

\item{burden}{Perform simple (unweighted) burden test? Must be TRUE or FALSE.}

\item{wburden}{Perform weight burden test? Weights are from beta(MAF, 1, 25)
distribution. Must be TRUE or FALSE.}

\item{SKAT}{Perform simple (unweighted) SKAT? Must be TRUE or FALSE.}

\item{altCovariancePath}{Option to give alternative path to write covariance
files to. Default = NULL. If left as NULL, covariance files will be written
in current directory.}

\item{altGroupFilePath}{Optional alterntive path to write group files to.
Default = NULL.}

\item{altRaremetalPath}{Optional alternative path to write RAREMETAL results
to. Default = NULL. If left as NULL, will write in current directory.}

\item{gene}{The name of the gene to use in the name of the output file.
Default = 'gene'.}

\item{hwe}{Hardy-Weinberg equilibrium p-value cut-off. Default = 0.000001.}

\item{altRaremetalName}{Optional command to call RAREMETAL with. For example if
you have multiple RAREMETAL installations and want to use a specific one, please
give the path + name of the executable file for the installation you want to use.
Default = NULL.}
}
\value{
Nothing. Will write the RAREMETAL results to
altRaremetalPath/gene.mask_name.meta.burden.results for burden,
altRaremetalPath/gene.mask_name.meta.BBeta.results for weighted burden,
and altRaremetalPath/gene.mask_name.meta.SKAT_.results for SKAT
}
\description{
A function that performs simple burden testing by making a call
to RAREMETAL
}
