% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covariance.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Covariance computation using external reference panel}
\usage{
compute_covariance(
  gt,
  allele_freq_test,
  n,
  residual_variance,
  altCovariancePath = NULL,
  gene = "gene"
)
}
\arguments{
\item{gt}{Sparse genotype matrix containing only the variants present in the
test allele frequency file (ie. bi-allelic variants w/ AF > 0).}

\item{allele_freq_test}{The in-sample allele frequency data frame with columns
CHROM, POS, REF, ALT, AF, AC, U_STAT, and PVAL.}

\item{n}{The sample size of the original study}

\item{residual_variance}{The residual variance from the original study}

\item{altCovariancePath}{Option to give alternative path to write covariance
files to. Default = NULL. If left as NULL, covariance files will be written
in current directory.}

\item{gene}{The name of the gene to use in the name of the covariance file.
Default = 'gene'.}
}
\value{
Nothing. Writes and bgzips covariance file in the format
altCovariancePath/gene.estimated.cov.gz
}
\description{
A function that computes covariance for rare-variant aggregation
testing using a VCF from an external reference panel and write it
}
