% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_frequency.R
\name{generate_allele_frequency}
\alias{generate_allele_frequency}
\title{Generate a data frame with allele frequencies from VCF}
\usage{
generate_allele_frequency(vcf_file)
}
\arguments{
\item{vcf_file}{String containing the file path and name containing the VCF
of the genotypes of the  set of individuals to be used as the reference
panel.}
}
\value{
A data frame with columns CHROM, POS, REF, ALT, AN, AC, AF
}
\description{
A function that takes a VCF file and returns a data frame containing
columns CHROM, POS, REF, ALT, AN, AC, AF
}
