% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_group_file.R
\name{generate_group_file}
\alias{generate_group_file}
\title{Generate group/mask file for rare-variant aggregation testing}
\usage{
generate_group_file(
  anno,
  allele_freq,
  gene,
  pLOF = TRUE,
  pLOF_narrowMissense = TRUE,
  pLOF_broadMissense = TRUE,
  altGroupFilePath = NULL,
  mafThreshold = 0.01
)
}
\arguments{
\item{anno}{A VEP annotation file with at minimum columns named #Uploaded_variation
containing the variants in format chr:pos:ref:alt, Consequence containing the predicted
consequences of the variants, SYMBOL containing the gene names that the variants
belong to, and PICK which is 1 if that row is to be used and "-" otherwise,
a column sumAlgs which is 1 if deleterious and 0 otherwise OR LRT_pred,
MutationTaster_pred, Polyphen2_HDIV_pred, Polyphen2_HVAR_pred, and
SIFT4G_pred containing LRT, MutationTaster, Polyphen2_HDIV, Polyphen2_HVAR,
and SIFT4G deleteriousness predictions respectively.}

\item{allele_freq}{The in-sample allele frequency data frame with columns
CHROM, POS, REF, ALT, AF, AC, U_STAT, and PVAL.}

\item{gene}{The name of the gene from the annotation file to use.}

\item{pLOF}{TRUE or FALSE. Perform aggregation tests using pLOF variants only?
Default = TRUE.}

\item{pLOF_narrowMissense}{TRUE or FALSE. Perform aggregation tests using
pLOF + missense(narrow) variants only? Default = TRUE.}

\item{pLOF_broadMissense}{TRUE or FALSE. Perform aggregation tests using
pLOF + missense(broad) variants only? Default = TRUE.}

\item{altGroupFilePath}{Optional alterntive path to write group files to.
Default = NULL.}

\item{mafThreshold}{Threshold for including variants in group file. Note here
we assume that the alternate allele is coded as the minor allele.
Default = 0.01.}
}
\value{
List of data frames containing the SNPs in each mask
}
\description{
A function that generates and writes to disk in RAREMETAL format
 a group/mask file for rare variant aggregation testing.
}
