% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_user_group_file.R
\name{read_user_group_file}
\alias{read_user_group_file}
\title{Read in user-supplied group files}
\usage{
read_user_group_file(group_file)
}
\arguments{
\item{group_file}{Vector of strings containing the file paths and names of
 the group files to perform the rare-variant aggregation tests with.

 @return A list where each element is a vector of the variants in the
 user-supplied mask.}
}
\description{
Read in user-supplied group files
}
