% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in a VCF to a sparse genotype matrix}
\usage{
read_vcf(
  vcf_file,
  chr,
  allele_freq_test,
  allele_freq_reference,
  original_snps,
  gene_start = NULL,
  gene_end = NULL
)
}
\arguments{
\item{vcf_file}{String containing the file path and name containing the VCF
of the genotypes of the  set of individuals to be used as the reference
panel. Note that the chromosome column must be in the format c{chr_number}.
ie. c1 for chromosome 1.}

\item{chr}{The chromosome number on which you want to perform the analysis.
Must be in the format c{chr_number} or chr{chr_number}. ie. c1 or chr1 for
chromosome 1.}

\item{allele_freq_test}{Data frame from test set VCF containing
the columns CHROM, POS, REF, ALT, AF, AC, U_STAT, PVAL, SNP}

\item{allele_freq_reference}{Data frame from reference panel VCF containing
the columns CHROM, POS, REF, ALT, AN, AC, AF, SNP}

\item{original_snps}{A vector containing the SNPs that were in the original
reference allele frequency file before filtering/matching to the test set}

\item{gene_start}{The starting base pair position of the gene on the chromosome.
Default = NULL.}

\item{gene_end}{The ending base pair position of the gene on the chromosome.
Default = NULL.}
}
\value{
A sparse genotype matrix with the genotypes of the individuals,
colnames corresponding to the SNPs in format CHR:REF:ALT:POS, and rownames
corresponding to the participant IDs
}
\description{
A function that takes a VCF file and returns a genotype matrix with
the genotypes of the individuals, colnames corresponding to the variant positions,
and rownames corresponding to the ID of the individuals. Only returns genotypes
of variants that are bi-allelic and have allele frequency > 0 in the test set.
ie. the ones that are present in the allele frequency files.
}
