% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_residual_variance.R
\name{ss_residual_variance}
\alias{ss_residual_variance}
\title{Read in sample size and residual variance from RAREMETALWORKER score
stat file.}
\usage{
ss_residual_variance(score_stat_file)
}
\arguments{
\item{score_stat_file}{String containing the file path and name containing
 the RAREMETALWORKER formatted score statistic file.

 @importFrom {data.table} {fread}}
}
\value{
A list containing the residual variance and sample size of the study
}
\description{
Read in sample size and residual variance from RAREMETALWORKER score
stat file.
}
