% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_test.R
\name{two_stage_test}
\alias{two_stage_test}
\title{Perform two-stage approach for rare-variant aggregation testing}
\usage{
two_stage_test(
  mask_list,
  allele_freq_test,
  two_stage_threshold = 3,
  residual_variance,
  sample_size
)
}
\arguments{
\item{mask_list}{A list where each element is a 1 column data frame of variants
to include in each test.}

\item{allele_freq_test}{Data frame containing the columns "CHROM", "POS",
"REF", "ALT", "AF", "AC", "PVAL"}

\item{two_stage_threshold}{-log10(p-value) threshold to use for two-stage
approach. Default = 3.}

\item{residual_variance}{The residual variance from the study}

\item{sample_size}{The sample size from the study}
}
\value{
TRUE/FALSE to compute covariance or not if at least one of the masks
passes the -log10(p-value) threshold
}
\description{
A function that performs the two-stage approach for rare-variant
aggregation testing.
}
