% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_raremetal.R
\name{call_raremetal}
\alias{call_raremetal}
\title{Burden testing using RAREMETAL}
\usage{
call_raremetal(
  mask_list,
  score_stat_file,
  group_file_names,
  altCovariancePath = NULL,
  altGroupFilePath = NULL,
  altRaremetalPath = NULL,
  gene = "gene",
  hwe = 1e-06
)
}
\arguments{
\item{mask_list}{Named list of vectors containing the variants to use in the
group files for burden testing.}

\item{score_stat_file}{String containing the file path and name containing
the RAREMETALWORKER formatted score statistic file.}

\item{group_file_names}{Vector of strings containing the file paths and names of
the group files to perform the rare-variant aggregation tests with.}

\item{altCovariancePath}{Option to give alternative path to write covariance
files to. Default = NULL. If left as NULL, covariance files will be written
in current directory.}

\item{altGroupFilePath}{Optional alterntive path to write group files to.
Default = NULL.}

\item{altRaremetalPath}{Optional alternative path to write RAREMETAL results
to. Default = NULL. If left as NULL, will write in current directory.}

\item{gene}{The name of the gene to use in the name of the output file.
Default = 'gene'.}

\item{hwe}{Hardy-Weinberg equilibrium p-value cut-off. Default = 0.000001.}
}
\value{
Nothing. Will write the RAREMETAL results to
altRaremetalPath/gene.mask_name.burden.res - ??
}
\description{
A function that performs simple burden testing by making a call
to RAREMETAL
}
