% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_frequency.R
\name{generate_allele_frequency}
\alias{generate_allele_frequency}
\title{Generate a data frame with allele frequencies from VCF}
\usage{
generate_allele_frequency(vcf_file, chr, gene_start, gene_end)
}
\arguments{
\item{vcf_file}{String containing the file path and name containing the VCF
of the genotypes of the  set of individuals to be used as the reference
panel. Note that the chromosome column will automaticaly be set to the chromosome
given to the function.}

\item{chr}{The chromosome number on which you want to perform the analysis.
Must be in the format c{chr_number}. ie. c1 for chromosome 1.}

\item{gene_start}{The starting base pair position of the gene on the chromosome.
Default = NULL.}

\item{gene_end}{The ending base pair position of the gene on the chromosome.
Default = NULL.}
}
\value{
A data frame with columns CHROM, POS, REF, ALT, AN, AC, AF
}
\description{
A function that takes a VCF file and returns a data frame containing
columns CHROM, POS, REF, ALT, AN, AC, AF
}
